Renamed leader to meshcop_leader
