+++ src/core/Makefile.am
@@ -190,8 +190,8 @@ SOURCES_COMMON                             = \
     meshcop/energy_scan_client.cpp           \
     meshcop/joiner.cpp                       \
     meshcop/joiner_router.cpp                \
-    meshcop/leader.cpp                       \
     meshcop/meshcop.cpp                      \
+    meshcop/meshcop_leader.cpp               \
     meshcop/meshcop_tlvs.cpp                 \
     meshcop/panid_query_client.cpp           \
     meshcop/timestamp.cpp                    \
@@ -387,8 +387,8 @@ HEADERS_COMMON                             = \
     meshcop/energy_scan_client.hpp           \
     meshcop/joiner.hpp                       \
     meshcop/joiner_router.hpp                \
-    meshcop/leader.hpp                       \
     meshcop/meshcop.hpp                      \
+    meshcop/meshcop_leader.hpp               \
     meshcop/meshcop_tlvs.hpp                 \
     meshcop/panid_query_client.hpp           \
     meshcop/timestamp.hpp                    \
